move stats.py
